Fix: Wireshark crashes when using "important" filter
The important filter assumes parse-arguments is enabled in tracee, and
filters by the parsed value of the ptrace event's request field.
If parse-arguments is disabled, the field is an int which causes some
sort of confusion with the explicitly registered request field (defined
as a string), which causes Wireshark to crash when applying a filter
that compares the field's value with a string.

This commit overrides the field type of the request, and adds a new
enriched field to the event, which contains the parsed request.
The "important" filter now uses this parsed field instead.